# api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "MIshmarHevratyAPI",
      "description": "API For Mishmar.",
      "version": "1.0"
    },
    "produces": ["application/json"],
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/users": {
            "post": {
                "x-swagger-router-controller": "User",
                "operationId": "addUser",
                "description": "Register a new user to the notification system",
                "parameters": [
                {
                    "name": "name",
                    "in": "formData",
                    "description": "the user name",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "id",
                    "in": "formData",
                    "description": "the user ID",
                    "required": true,
                    "type": "string"
                }
                ],
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "user registered successfuly.",
                    "schema": {
                    "$ref": "#/definitions/Success"
                    }
                }
                }
            }
        },
        "/subjects": {
            "post": {
                "x-swagger-router-controller": "Subject",
                "operationId": "addSubject",
                "description": "Add a subject",
                "parameters": [
                {
                    "name": "subject",
                    "in": "formData",
                    "description": "Subject text",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "default": {
                        "description": "Invalid request.",
                        "schema": {
                        "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "subjects added successfuly.",
                        "schema": {
                        "$ref": "#/definitions/Success"
                        }
                    }
                }
            },
            "delete": {
                "x-swagger-router-controller": "Subject",
                "operationId": "deleteSubject",
                "description": "Delete a subject",
                "parameters": [
                {
                    "name": "subject",
                    "in": "formData",
                    "description": "Subject text",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "default": {
                        "description": "Invalid request.",
                        "schema": {
                        "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "subjects deleted successfuly.",
                        "schema": {
                        "$ref": "#/definitions/Success"
                        }
                    }
                }
            },
            "get": {
                "x-swagger-router-controller": "Subject",
                "operationId": "getSubjects",
                "description": "get all subjects",
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "All possible subjects.",
                    "schema": {
                    "$ref": "#/definitions/Subjects"
                    }
                }
                }
            }
        },       
        
        "/user_subjects": {
            "post": {
                "x-swagger-router-controller": "User",
                "operationId": "setSubjects",
                "description": "set subjects to a user",
                "parameters": [
                {
                    "name": "subjects",
                    "in": "formData",
                    "description": "subjects to register to the user",
                    "required": true,
                    "type": "array",
                    "items":{
                        "type": "string"
                    }
                }
                ],
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "subjects added to user successfuly.",
                    "schema": {
                    "$ref": "#/definitions/Success"
                    }
                }
                }
            },
            "get": {
                "x-swagger-router-controller": "User",
                "operationId": "getSubjects",
                "description": "get subjects user is registred",
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "subjects the user is registred to.",
                    "schema": {
                    "$ref": "#/definitions/Subjects"
                    }
                }
                }
            }
        },
        "/remarks": {
            "post": {
                "x-swagger-router-controller": "UserActions",
                "operationId": "addRemark",
                "description": "add a user remark to the commitee",
                "parameters": [
                {
                    "name": "committee_id",
                    "in": "formData",
                    "description": "commitee ID",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "remark",
                    "in": "formData",
                    "description": "user remark on the commitee",
                    "required": true,
                    "type": "string"
                }
                ],
                
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "remark loged successfuly.",
                    "schema": {
                    "$ref": "#/definitions/Success"
                    }
                }
                }
            },
            "get": {
                "x-swagger-router-controller": "UserActions",
                "operationId": "getRemarks",
                "description": "get session remarks",
                "parameters": [
                {
                    "name": "committee_id",
                    "in": "query",
                    "description": "commitee ID",
                    "required": true,
                    "type": "string"
                }
                ],
                "responses": {
                    "default": {
                        "description": "Invalid request.",
                        "schema": {
                        "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "remarks array of the commitee session.",
                        "schema": {
                        "$ref": "#/definitions/Remarks"
                        }
                    }
                }
            }
        },
        "/session_subjects": {
            "get": {
                "x-swagger-router-controller": "SessionSubject",
                "operationId": "getSubjects",
                "description": "get subjects of a session",
                "parameters": [
                {
                    "name": "session_id",
                    "in": "query",
                    "description": "retrieve the subjects for a session",
                    "required": true,
                    "type": "string"
                }
                ],
                "responses": {
                    "default": {
                        "description": "Invalid request.",
                        "schema": {
                        "$ref": "#/definitions/Error"
                        }
                    },
                    "200": {
                        "description": "sessions list.",
                        "schema": {
                        "$ref": "#/definitions/Subjects"
                        }
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "SessionSubject",
                "operationId": "addSessionSubject",
                "description": "add session and subject connection",
                "parameters": [
                {
                    "name": "subject",
                    "description": "subject",
                    "in": "formData",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "session_id",
                    "description": "id of the session",
                    "in": "formData",
                    "required": true,
                    "type": "string"
                }
                ],
                
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "remark loged successfuly.",
                    "schema": {
                    "$ref": "#/definitions/Success"
                    }
                }
                }
            },
            "delete": {
                "x-swagger-router-controller": "SessionSubject",
                "operationId": "deleteSessionSubject",
                "description": "remove session and subject connection",
                "parameters": [
                {
                    "in": "formData",
                    "name": "subject",
                    "description": "subject",
                    "required": true,
                    "type": "string"
                },
                {
                    "in": "formData",
                    "name": "session_id",
                    "description": "id of the session",
                    "required": true,
                    "type": "string"
                }
                ],
                
                "responses": {
                "default": {
                    "description": "Invalid request.",
                    "schema": {
                    "$ref": "#/definitions/Error"
                    }
                },
                "200": {
                    "description": "Session/subject link removed successfuly.",
                    "schema": {
                        "$ref": "#/definitions/Success"
                    }
                }
                }
            }
        }
    },
    "definitions": {
      "Subjects": {
        "properties": {
            "subjects": {
                "type": "array",
                "items":{
                    "type": "string"
                }
            }
          },
          "required": ["subjects"]
      },
      "Sessions": {
        "properties": {
            "sessions": {
                "type": "array",
                "items":{
                    "type": "string"
                }
            }
          },
          "required": ["sessions"]
      },
      "Remarks": {
        "properties": {
            "remarks": {
                "type": "array",
                "items":{
                    "type": "string"
                }
            }
          },
          "required": ["remarks"]
      },
      "Error": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
      "Success": {
        "properties": {
            "success": {
              "type": "string"
            }
          },
          "required": ["success"]
      }
    }
}